requirements.txt
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dropbox==10.3.0
et-xmlfile==1.0.1
Flask==1.1.2
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.4.0
openpyxl==3.0.4
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.0.0
requests==2.24.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
uWSGI==2.0.19.1
Werkzeug==1.0.1
zipp==3.1.0
